+++ docs/cmd/tkn_pipelinerun_logs.md
@@ -32,6 +32,7 @@ Show the logs of PipelineRun named 'microservice-1' for all tasks and steps (inc
 ```
   -a, --all            show all logs including init steps injected by tekton
   -f, --follow         stream live logs
+  -F, --fzf            use fzf to select a pipelinerun
   -h, --help           help for logs
   -L, --last           show logs for last pipelinerun
       --limit int      lists number of pipelineruns (default 5)
