+++ docs/man/man1/tkn-pipelinerun-logs.1
@@ -28,6 +28,10 @@ Show the logs of PipelineRun
     stream live logs
 
 .PP
+\fB\-F\fP, \fB\-\-fzf\fP[=false]
+    use fzf to select a pipelinerun
+
+.PP
 \fB\-h\fP, \fB\-\-help\fP[=false]
     help for logs
