+++ docs/man/man1/tkn-taskrun-logs.1
@@ -28,6 +28,10 @@ Show taskruns logs
     stream live logs
 
 .PP
+\fB\-F\fP, \fB\-\-fzf\fP[=false]
+    use fzf to select a taskrun
+
+.PP
 \fB\-h\fP, \fB\-\-help\fP[=false]
     help for logs
