+++ pkg/cmd/pipelinerun/testdata/TestPipelineRunDescribe_cancelled_pipelinerun.golden
@@ -9,7 +9,7 @@ STARTED          DURATION    STATUS
 
 Message
 
-PipelineRun "pipeline-run" was cancelled
+PipelineRun &#34;pipeline-run&#34; was cancelled
 
 Resources
